Restructure auth
Anchor restructure
