+++ wallet-sdk/src/test/kotlin/org/stellar/walletsdk/AuthTest.kt
@@ -14,12 +14,11 @@ internal class AuthTest : SuspendTest() {
   fun `auth token`() {
     val authToken = runBlocking {
       Auth(
-          accountAddress = ADDRESS_ACTIVE,
           webAuthEndpoint = AUTH_ENDPOINT,
           homeDomain = AUTH_HOME_DOMAIN,
           walletSigner = InProcessWalletSigner()
         )
-        .authenticate()
+        .authenticate(ADDRESS_ACTIVE)
     }
 
     assertNotNull(authToken)
@@ -28,14 +27,8 @@ internal class AuthTest : SuspendTest() {
   @Test
   fun `auth token with client domain`() {
     val authToken = runBlocking {
-      Auth(
-          ADDRESS_ACTIVE,
-          AUTH_ENDPOINT,
-          AUTH_HOME_DOMAIN,
-          clientDomain = AUTH_CLIENT_DOMAIN,
-          walletSigner = InProcessWalletSigner()
-        )
-        .authenticate()
+      Auth(AUTH_ENDPOINT, AUTH_HOME_DOMAIN, walletSigner = InProcessWalletSigner())
+        .authenticate(ADDRESS_ACTIVE, clientDomain = AUTH_CLIENT_DOMAIN)
     }
 
     assertNotNull(authToken)
@@ -45,15 +38,8 @@ internal class AuthTest : SuspendTest() {
   fun `throw exception if both memo and clientDomain are provided`() {
     assertThrows<Exception> {
       runBlocking {
-        Auth(
-            ADDRESS_ACTIVE,
-            AUTH_ENDPOINT,
-            AUTH_HOME_DOMAIN,
-            memoId = "123",
-            clientDomain = AUTH_CLIENT_DOMAIN,
-            walletSigner = InProcessWalletSigner()
-          )
-          .authenticate()
+        Auth(AUTH_ENDPOINT, AUTH_HOME_DOMAIN, walletSigner = InProcessWalletSigner())
+          .authenticate(ADDRESS_ACTIVE, memoId = "123", clientDomain = AUTH_CLIENT_DOMAIN)
       }
     }
   }
@@ -62,14 +48,8 @@ internal class AuthTest : SuspendTest() {
   fun `throw exception if Memo ID is not a positive integer`() {
     assertThrows<Exception> {
       runBlocking {
-        Auth(
-            ADDRESS_ACTIVE,
-            AUTH_ENDPOINT,
-            AUTH_HOME_DOMAIN,
-            memoId = "abc",
-            walletSigner = InProcessWalletSigner()
-          )
-          .authenticate()
+        Auth(AUTH_ENDPOINT, AUTH_HOME_DOMAIN, walletSigner = InProcessWalletSigner())
+          .authenticate(ADDRESS_ACTIVE, memoId = "abc")
       }
     }
   }
